requirements.txt
numpy
opencv-python
pyyaml
networkx
matplotlib
Pillow
scikit-image
scipy
tqdm
faiss-cpu
qdrant-client
sentence-transformers
transformers
torch
torchvision
accelerate
safetensors
huggingface-hub
